refactored ga to modded_pygad_ga to avoid confusion
